Simplify user edit logic, fix API key regeneration
The logic for allowing local edits on user accounts was extremely
complex and hard to follow. It allow individual users to choose whether
they wanted to update their details but it was fairly broken.
A specially significant broke feature was the ability to regenerate API
keys, as that uses a separate endpoint that was not handling the custom
fields implemented by this extension.

This change implements a site-wide policy for local user edits based
solely on the `saml2.allow_user_changes` config option (renamed from
`ckan.saml2.allow_user_changes` for consistency) and drastically reduces
the logic.